requirements.txt
boto3
python-dotenv
langsmith
langchain
openai
supabase